Critical: Filter out iPods in Makefile (#597)
This small, but pretty significant change filters iPods in the Makefile, preventing Procursus from using incorrect variables.

What happens is that, prior to this change, Procursus would "report" that packages were being built on macOS on iPods, despite that not being true. This fixes the issue, meaning that iPods are now recognized as an iOS platform, and the correct iOS variables are now used.